updated publications
