+++ .gitignore
@@ -39,3 +39,4 @@ publications/WHO Nepal Paper/.DS_Store
 publications/PISA Timor Leste/.ipynb_checkpoints/Timor Leste_Estrada_WorldPop_Analysis-checkpoint.ipynb
 publications/PISA Timor Leste/.DS_Store
 *.pkl
+.DS_Store
