--- publications/WHO Nepal Paper/Data/readme.md
@@ -1,1 +1,0 @@
-
